"Kleinere Anpassungen."
git-svn-id: svn://svn.berlios.de/dsa-hl@1875 4c79e8ff-cfd4-0310-af45-a38c79f83013

+++ rl/trunk/engine/multimedia/src/fmod3/Fmod3Driver.cpp
@@ -478,7 +478,7 @@ void Fmod3Driver::printData()
 void Fmod3Driver::setMasterVolume(unsigned int vol)
 {
 	SoundDriver::setMasterVolume(vol);
-	FSOUND_SetSFXMasterVolume(vol * 2.55);
+	FSOUND_SetSFXMasterVolume(int(vol * 2.55));
 }
